added delete incident, reorganized files
